# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  codeberg_org_azzet_azzetbe_internal_claim.ClaimDetailResponse:
    properties:
      audit_log:
        items:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.AuditLogEntry'
        type: array
      claimant_entity_id:
        type: string
      claimant_user_id:
        type: string
      created_at:
        type: string
      dispute_reason:
        type: string
      documents:
        items:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.DocumentResponse'
        type: array
      entity_id:
        type: string
      entity_name:
        type: string
      entity_type:
        type: string
      id:
        type: string
      notes:
        type: string
      rejection_reason:
        type: string
      reviewed_at:
        type: string
      reviewer_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_claim.ClaimListResponse:
    properties:
      claimant_name:
        type: string
      created_at:
        type: string
      document_count:
        type: integer
      entity_id:
        type: string
      entity_name:
        type: string
      entity_type:
        type: string
      id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  codeberg_org_azzet_azzetbe_internal_claim.DisputeRequest:
    properties:
      reason:
        type: string
    type: object
...
  codeberg_org_azzet_azzetbe_internal_shared.APIResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
  codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIError'
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Enterprise-grade accounting, tax, and finance platform. User routes
    use Bearer JWT from /auth/login. Workspace-scoped routes require X-Workspace-ID
    header. Admin routes under /admin use admin JWT from /admin/auth/login.
  title: Azzet API
  version: "1.0"
paths:
...
  /admin/claims:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.ClaimListResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List claims for admin review
      tags:
      - Admin Claim
  /admin/claims/{id}:
    get:
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.ClaimDetailResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get claim detail (admin)
      tags:
      - Admin Claim
  /admin/claims/{id}/approve:
    post:
      consumes:
      - application/json
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Reviewer notes
        in: body
        name: body
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve claim
      tags:
      - Admin Claim
  /admin/claims/{id}/assign:
    post:
      description: Moves claim from SUBMITTED/DISPUTED to UNDER_REVIEW
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign claim to reviewer
      tags:
      - Admin Claim
  /admin/claims/{id}/audit:
    get:
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.AuditLogEntry'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get claim audit log
      tags:
      - Admin Claim
  /admin/claims/{id}/documents/{doc_id}/view:
    get:
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Document ID
        in: path
        name: doc_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Get presigned URL to view claim document
      tags:
      - Admin Claim
  /admin/claims/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason required
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject claim
      tags:
      - Admin Claim
  /admin/claims/stats:
    get:
      description: Returns count of claims in SUBMITTED status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Count pending claims
      tags:
      - Admin Claim
...
  /claims:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.ClaimListResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List my claims
      tags:
      - Claim
    post:
      consumes:
      - application/json
      description: Start a claim workflow for a shadow entity
      parameters:
      - description: Claim request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.CreateClaimRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.ClaimResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a company claim
      tags:
      - Claim
  /claims/{id}:
    get:
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.ClaimDetailResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get claim detail
      tags:
      - Claim
  /claims/{id}/dispute:
    post:
      consumes:
      - application/json
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Dispute reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.DisputeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dispute a rejected claim
      tags:
      - Claim
  /claims/{id}/documents:
    get:
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.DocumentResponse'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List claim documents
      tags:
      - Claim
    post:
      consumes:
      - application/json
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Upload request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.DocumentUploadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.PresignedUploadResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request claim document upload URL
      tags:
      - Claim
  /claims/{id}/documents/{doc_id}/confirm:
    post:
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Document ID
        in: path
        name: doc_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm claim document upload
      tags:
      - Claim
  /claims/{id}/submit:
    post:
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_claim.ClaimResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit claim for review
      tags:
      - Claim
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"